add generic provider
