+++ provider_container_settings.json
@@ -1,12 +1,13 @@
 [
     {
         "name": "go",
-        "binaryPath": "/usr/bin/golang-external-provider",
+        "binaryPath": "/usr/bin/generic-external-provider",
         "initConfig": [{
             "location": "examples/golang",
             "analysisMode": "full",
             "providerSpecificConfig": {
-                "lspServerPath": "/root/go/bin/gopls"
+                "lspServerPath": "/root/go/bin/gopls",
+                "dependencyProviderPath": "golang-dependency-provider"
             }
         }]
     },
